sel4arm-vmm: fix types for 64-bit
Use seL4_Word, long, uintptr_t in place of uint32_t.

+++ libsel4arm-vmm/src/copyinout.c
@@ -146,7 +146,7 @@ vm_copyout(vm_t* vm, void* data, uintptr_t address, size_t size)
 void*
 vm_copyout_elf(vm_t* vm, void* elf_file)
 {
-    uint32_t entry;
+    long entry;
     int num_headers;
     int i;
